app/requirements.txt

spacy
pdfplumber
sentence-transformers
faiss-cpu
requests
